% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bootstrap_LRT_mclust}
\alias{bootstrap_LRT_mclust}
\title{Bootstrap the likelihood-ratio test statistic for mixture components}
\usage{
bootstrap_LRT_mclust(df, model_names, ...)
}
\arguments{
\item{df}{data.frame with two or more columns with continuous variables}

\item{model_names}{names of one or more models ?run mclust::mclustModelNames() to see all of the possible models and their names / abbreviations)}

\item{...}{additional arguments to be passed to mclustBootstrapLRT()}
}
\description{
Bootstrap the likelihood-ratio test statistic for mixture components
}
\details{
Bootstrap the p-values for the likelihood-ratio test statistic for the number of mixture components for an mclust model
}
