% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_mclust.R
\name{calculate_centroids_mclust}
\alias{calculate_centroids_mclust}
\title{Calculate centroids from an mclust model object}
\usage{
calculate_centroids_mclust(x)
}
\arguments{
\item{x}{an object of class `Mclust`}
}
\description{
Calculate centroids from an mclust model object
}
\details{
Extract the centroid values of an mclust model object
}
