% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models_lpa.R
\name{compare_models_lpa}
\alias{compare_models_lpa}
\title{Explore BIC of mclust models}
\usage{
compare_models_lpa(df, ..., n_profiles_range = 1:9, model = c(1, 2, 3),
  statistic = "BIC", return_table = FALSE)
}
\arguments{
\item{df}{data.frame with two or more columns with continuous variables}

\item{...}{unquoted variable names separated by commas}

\item{n_profiles_range}{a vector with the range of the number of mixture components to explore; defaults to 1 through 9 (1:9)}

\item{model}{the mclust model to explore: 1 (varying means, equal variances, and residual covariances fixed to zero); 2 (varying means, equal variances and covariances; and 3 (varying means, variances, and covariances), in order from most to least constrained; run ?mclust::mclustModelNames() to see all of the possible models and their names / abbreviations)}

\item{statistic}{what statistic to plot; BIC or ICL are presently available as options}

\item{return_table}{logical (TRUE or FALSE) for whether to return a table of the output intsead of a plot; defaults to FALSE}
}
\value{
a ggplot2 plot of the BIC values for the explored models
}
\description{
Explore BIC of mclust models
}
\details{
Explore the BIC values of a range of models in terms of a) the structure of the residual covariance matrix and b) the number of mixture components (or profiles)
}
\examples{
d <- iris
compare_models_lpa(d, Sepal.Length, Sepal.Width, Petal.Length, Petal.Width)
}
