% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models_mplus.R
\name{compare_models_mplus}
\alias{compare_models_mplus}
\title{Explore BIC of MPlus models}
\usage{
compare_models_mplus(df, ..., n_profiles = 2:10, models = 1:5,
  start_iterations = c(100, 20), m_iterations = 500, save_models = NULL)
}
\arguments{
\item{df}{data.frame with two or more columns with continuous variables}

\item{...}{unquoted variable names separated by commas}

\item{n_profiles_range}{a vector with the range of the number of mixture components to explore; defaults to 2 through 10 (2:10)}
}
\value{
a list with a data.frame with the BIC values and a list with all of the model output; if save_models is the name of an rds file (i.e., "out.rds"), then the model output will be written with that filename and only the data.frame will be returned
}
\description{
Explore BIC of MPlus models
}
\details{
Explore the BIC values of a range of Mplus models in terms of a) the structure of the residual covariance matrix and b) the number of mixture components (or profiles)
}
\examples{
compare_models_mplus(iris, Sepal.Length, Sepal.Width, Petal.Length, Petal.Width)
}
