% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profiles_lpa.R
\name{create_profiles_lpa}
\alias{create_profiles_lpa}
\title{Create profiles for a specific mclust model}
\usage{
create_profiles_lpa(df, ..., n_profiles, model = 1, to_return = "tibble",
  return_posterior_probs = TRUE)
}
\arguments{
\item{df}{data.frame with two or more columns with continuous variables}

\item{...}{unquoted variable names separated by commas}

\item{n_profiles}{the number of profiles (or mixture components) to be estimated}

\item{model}{the mclust model to explore: 1 (varying means, equal variances, and residual covariances fixed to zero); 2 (varying means, equal variances and covariances; and 3 (varying means, variances, and covariances), in order from most to least constrained; run ?mclust::mclustModelNames() to see all of the possible models and their names / abbreviations)}

\item{to_return}{character string for either "tibble" or "mclust" if "tibble" is selected, then data with a column for profiles is returned; if "mclust" is selected, then output of class mclust is returned}

\item{return_posterior_probs}{TRUE or FALSE (only applicable if to_return == "tibble"); whether to include posterior probabilities in addition to the posterior profile classification; defaults to FALSE}
}
\value{
either a tibble or a ggplot2 plot of the BIC values for the explored models
}
\description{
Create profiles for a specific mclust model
}
\details{
Creates profiles (or estimates of the mixture components) for a specific mclust model in terms of the specific number of mixture components and the structure of the residual covariance matrix
}
\examples{
\dontrun{
d <- pisaUSA15
m3 <- create_profiles_lpa(d,
                             broad_interest, enjoyment, instrumental_mot, self_efficacy,
                             n_profiles = 3, to_return="tibble")
}
}
