% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_profiles_mclust}
\alias{create_profiles_mclust}
\title{Create profiles for a specific mclust model}
\usage{
create_profiles_mclust(df, n_profiles, variance_structure = "freed",
  covariance_structure = "freed", model_name = NULL)
}
\arguments{
\item{df}{data.frame with two or more columns with continuous variables}

\item{n_profiles}{the number of profiles (or mixture components) to be estimated}

\item{variance_structure}{specification for residual variances, either "constrained" (to be equivalent across profiles) or "freed" (to be freely-estimated across profiles)}

\item{covariance_structure}{specification for residual variance, either "fixed" (to be zero inall profiles), "constrained" (to be equivalent across profiles), or "freed" (to be freely-estimated across profiles)}

\item{model_name}{the mclust model to explore, such as constrained variance, fixed variances ("EII"), constrained variance, constrained covariance ("EEE"), and freed variance, freed covariance ("VVV"); or others (run mclust::mclustModelNames() to see all of the possible models and their names / abbreviations); if specified, arguments variance_structure and covariance_structure are ignored}
}
\value{
a ggplot2 plot of the BIC values for the explored models
}
\description{
Create profiles for a specific mclust model
}
\details{
Creates profiles (or estimates the mixture components) for a specific mclust model in terms of the specific number of mixture components and the structure of the residual covariance matrix
}
\examples{
library(dplyr)
df <- select(iris, -Species)
create_profiles_mclust(df, n_profiles = 3, variance_structure = "freed", covariance_structure = "freed")
}
