% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-profiles-mclust.R
\name{estimate_profiles_mclust}
\alias{estimate_profiles_mclust}
\title{Estimate latent profiles using mclust}
\usage{
estimate_profiles_mclust(df, n_profiles, model_numbers, ...)
}
\arguments{
\item{df}{data.frame with two or more columns with continuous variables}

\item{n_profiles}{Numeric vector. The number of profiles (or mixture
components) to be estimated. Each number in the vector corresponds to an
analysis with that many mixture components.}

\item{model_numbers}{Numeric vector. Numbers of the models to be estimated.
See \code{\link{estimate_profiles}} for a description of the models available
in tidyLPA.}

\item{...}{Parameters passed directly to \code{\link[mclust]{Mclust}}. See
the documentation of \code{\link[mclust]{Mclust}}.}
}
\value{
An object of class 'tidyLPA' and 'list'
}
\description{
Estimates latent profiles (finite mixture models) using the open source
package \code{\link[mclust:Mclust]{mclust}}.
}
\author{
Caspar J. van Lissa
}
