% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_mclust.R
\name{extract_covariance}
\alias{extract_covariance}
\title{Extract mclust covariances}
\usage{
extract_covariance(x, profile_n)
}
\arguments{
\item{x}{object of class Mclust}
}
\description{
Extract mclust covariances
}
\details{
Extract the covariances
}
