% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_LL_mplus}
\alias{extract_LL_mplus}
\title{Extract log-likelihoods from models fit with estimate_profiles_mplus()}
\usage{
extract_LL_mplus(output_filename = "i.out")
}
\arguments{
\item{output_filename}{name of output_filename from estimate_profiles_mplus()}
}
\value{
a tibble or a ggplot2 plot of the BIC values for the explored modelswith the log-likelihood, random start seed, and the number of the iteration
}
\description{
Extract log-likelihoods from models fit with estimate_profiles_mplus()
}
\details{
Extract log-likelihoods associated with solutions from random starts from estimate_profiles_mplus(). Note that return_tmp_files = FALSE must be added to estimate_profiles_mplus() for this function to work.
}
\examples{
\dontrun{
m1 <- estimate_profiles_mplus(iris,
                            Sepal.Length, Sepal.Width, Petal.Length, Petal.Width,
                            n_profiles = 2,
                            remove_tmp_files = FALSE)
extract_LL_mplus()
}
}
