% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_mclust_classifications}
\alias{extract_mclust_classifications}
\title{Extract mclust classifications from the function create_profiles_mclust()}
\usage{
extract_mclust_classifications(x)
}
\arguments{
\item{x}{an object of class `Mclust`}
}
\description{
Extract mclust classifications from the function create_profiles_mclust()
}
\details{
Extract the classifications, in the form of posterior probabilties, for specific observations from the mclust output from the function create_profiles_mclust() so that posterior probabilities for specific observations, statistics related to the estimation, and other output can be viewed
}
