% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_mclust_output}
\alias{extract_mclust_output}
\title{Extract mclust output from the function create_profiles()}
\usage{
extract_mclust_output(x)
}
\arguments{
\item{x}{an object of class `Mclust`}
}
\description{
Extract mclust output from the function create_profiles()
}
\details{
Extract the output of the mclust output from the function create_profiles() so that posterior probabilities for specific observations, statistics related to the estimation, and other output can be viewed
}
