% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_mclust.R
\name{extract_mclust_summary}
\alias{extract_mclust_summary}
\title{Extract mclust summary statistics}
\usage{
extract_mclust_summary(x)
}
\arguments{
\item{x}{object of class Mclust}
}
\description{
Extract mclust summary statistics
}
\details{
Extract the log likelihood, BIC, and entropy statistics
}
\examples{
d <- pisaUSA15
d <- sample_n(pisaUSA15, 500)
m <- create_profiles_lpa(d, broad_interest, enjoyment, self_efficacy, n_profiles = 4, to_return = "mclust")
extract_mclust_summary(m)
}
