% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_means}
\alias{extract_means}
\title{Extract mclust means}
\usage{
extract_means(x)
}
\arguments{
\item{x}{an object of class `Mclust`}
}
\description{
Extract mclust means
}
\details{
Extract the means
}
