% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_profiles_mplus.R
\name{extract_mplus_summary}
\alias{extract_mplus_summary}
\title{Extract summary statistics from an Mplus model}
\usage{
extract_mplus_summary(x)
}
\arguments{
\item{x}{an mplus model}
}
\value{
a tibble with summary statistics
}
\description{
Extract summary statistics from an Mplus model
}
\details{
Extract log likelihood, BIC, and entropy statistics from an Mplus model
}
