% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_edu.R
\docType{data}
\name{id_edu}
\alias{id_edu}
\title{Simulated identity data}
\format{A data frame with 443 rows and 16 variables.}
\usage{
data(id_edu)
}
\description{
This simulated dataset, based on Crochetti et al., 2014, contains five annual
assessments of adolescents' mean scores on the commitment, exploration (in
depth), and reconsideration subscales of the Utrecht-Management of Identity
Commitments Scale (Crocetti et al., 2008). The scores reported here reflect
the educational identity subscales of this instrument. The first measurement
wave occurred when adolescents were, on
average, 14 years old, and the last one when they were 18 years old.
}
\details{
\tabular{lll}{
  \strong{commitment1} \tab \code{numeric} \tab Mean score
  of educational commitment in
  wave 1\cr
  \strong{exploration1} \tab \code{numeric} \tab Mean score
  of educational exploration in
  wave 1\cr
  \strong{reconsideration1} \tab \code{numeric} \tab Mean score
  of educational reconsideration in
  wave 1\cr
  \strong{commitment2} \tab \code{numeric} \tab Mean score
  of educational commitment in
  wave 2\cr
  \strong{exploration2} \tab \code{numeric} \tab Mean score
  of educational exploration in
  wave 2\cr
  \strong{reconsideration2} \tab \code{numeric} \tab Mean score
  of educational reconsideration in
  wave 2\cr
  \strong{commitment3} \tab \code{numeric} \tab Mean score
  of educational commitment in
  wave 3\cr
  \strong{exploration3} \tab \code{numeric} \tab Mean score
  of educational exploration in
  wave 3\cr
  \strong{reconsideration3} \tab \code{numeric} \tab Mean score
  of educational reconsideration in
  wave 3\cr
  \strong{commitment4} \tab \code{numeric} \tab Mean score
  of educational commitment in
  wave 4\cr
  \strong{exploration4} \tab \code{numeric} \tab Mean score
  of educational exploration in
  wave 4\cr
  \strong{reconsideration4} \tab \code{numeric} \tab Mean score
  of educational reconsideration in
  wave 4\cr
  \strong{commitment5} \tab \code{numeric} \tab Mean score
  of educational commitment in
  wave 5\cr
  \strong{exploration5} \tab \code{numeric} \tab Mean score
  of educational exploration in
  wave 5\cr
  \strong{reconsideration5} \tab \code{numeric} \tab Mean score
  of educational reconsideration in
  wave 5\cr
  \strong{sex} \tab \code{factor} \tab Adolescent sex; M = male, F = female.
}
}
\references{
Crocetti, E., Klimstra, T. A., Hale, W. W., Koot, H. M., &
Meeus, W. (2013). Impact of early adolescent externalizing problem behaviors
on identity development in middle to late adolescence: A prospective 7-year
longitudinal study. Journal of Youth and Adolescence, 42(11), 1745-1758.
\href{https://doi.org/10.1007/s10964-013-9924-6}{
DOI:10.1007/s10964-013-9924-6}
}
\keyword{datasets}
