% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{An object and a function to apply to it}
}
\description{
tidyLPA suggests using the pipe operator, \code{\%>\%}, from the magrittr
package (imported here from the dplyr package).
}
\examples{
# Instead of
subset(iris, select = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"))
# you can write
iris \%>\%
  subset(select = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"))
}
