% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{pisaUSA15}
\alias{pisaUSA15}
\title{student questionnaire data with four variables from the 2015 PISA for students in the United States}
\format{Data frame with columns
#' \describe{
  \item{broad_interest}{composite measure of students' self reported broad interest}
  \item{enjoyment}{composite measure of students' self reported enjoyment}
  \item{instrumental_mot}{composite measure of students' self reported instrumental motivation}
  \item{self_efficacy}{composite measure of students' self reported self efficacy}
  ...
}}
\source{
http://www.oecd.org/pisa/data/
}
\usage{
pisaUSA15
}
\description{
student questionnaire data with four variables from the 2015 PISA for students in the United States
}
\keyword{datasets}
