% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tidyLPA.R
\name{plot.tidyLPA}
\alias{plot.tidyLPA}
\title{Plot tidyLPA object}
\usage{
\method{plot}{tidyLPA}(x, y, ..., statistics = "BIC")
}
\arguments{
\item{x}{an object used to select a method.}

\item{y}{Ignored.}

\item{...}{further arguments passed to or from other methods.}

\item{statistics}{Character vector, indicating which statistics to plot.
Defaults to "BIC".}
}
\description{
Plot tidyLPA object
}
