% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_mclust}
\alias{plot_mclust}
\title{Plot mclust centroids}
\usage{
plot_mclust(x)
}
\arguments{
\item{x}{an object of class `Mclust`}
}
\description{
Plot mclust centroids
}
\details{
Plot the centroids for Mclust output
}
