% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_profiles_mclust}
\alias{plot_profiles_mclust}
\title{Plot profile centroids}
\usage{
plot_profiles_mclust(d, to_center = F, to_scale = F)
}
\arguments{
\item{d}{output from create_profiles_mclust()}

\item{to_center}{whether to center the data before plotting}

\item{to_scale}{whether to scale the data before plotting}
}
\description{
Plot profile centroids
}
\details{
Plot the centroids for Mclust output
}
