% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{plot_profiles_mplus-deprecated}
\alias{plot_profiles_mplus-deprecated}
\title{Plot variable means and their confidence intervals by profile for models estimated with MPlus (requires purchasing and installing MPlus to use)}
\usage{
<<<<<<< HEAD:man/plot_profiles_mplus.Rd
plot_profiles_mplus(mplus_data = NULL, to_center = FALSE,
  to_scale = FALSE, plot_post_probs = FALSE, mplus_out_name = NULL,
  standard_error_interval = 0.95)
=======
plot_profiles_mplus(...)
>>>>>>> ca12dc048a1f699748149baabb5d1be33e1365bc:man/plot_profiles_mplus-deprecated.Rd
}
\arguments{
\item{x}{output from estimate_profiles()}

\item{mplus_data}{output from estimate_profiles_mplus() with return_savedata = T specified}

\item{plot_post_probs}{logical; whether to plot the means and standard errors based on the posterior probabilities (if TRUE) or on the classification/profile (if FALSE); defaults to TRUE}

\item{mplus_out_name}{character string; name of the mplus out file that is read if the posterior probabilities are plotted; defaults to i.out (which is the default name for the file created by other tidyLPA functions)}

\item{standard_error_interval}{number between 0 and 1; defaults to .95}

\item{to_center}{whether to center the data before plotting}

\item{to_scale}{whether to scale the data before plotting}

\item{plot_what}{whether to plot tibble or mclust output from estimate_profiles(); defaults to tibble}

\item{plot_error_bars}{whether to plot error bars (representing the 95 percent confidence interval for the mean of each variable)}

\item{plot_rawdata}{whether to plot raw data; defaults to TRUE}

\item{ci}{confidence interval to plot (defaults to 0.95)}
}
\description{
Plot variable means and their confidence intervals by profile for models estimated with MPlus (requires purchasing and installing MPlus to use)
}
\details{
Plot the variable means and variances for data frame output from estimate_profiles_mclust()
}
<<<<<<< HEAD:man/plot_profiles_mplus.Rd
\examples{
\dontrun{

m <- estimate_profiles_mplus(iris,
Sepal.Length, Sepal.Width, Petal.Length, Petal.Width,
n_profiles = 2)

plot_profiles_mplus(m)

m <- estimate_profiles_mplus(iris,
Sepal.Length, Sepal.Width, Petal.Length, Petal.Width,
n_profiles = 2, latent_vars = list(sepal = c(1, 2), petal = c(3, 4)),
remove_tmp_files = FALSE)

plot_profiles_mplus(mplus_out_name = "i.out")
}
}
=======
\seealso{
\code{\link{tidyLPA-deprecated}}
}
\keyword{internal}
>>>>>>> ca12dc048a1f699748149baabb5d1be33e1365bc:man/plot_profiles_mplus-deprecated.Rd
