% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-profiles.R
\name{plotMixtures}
\alias{plotMixtures}
\alias{plot_profiles}
\alias{plotMixtures.default}
\alias{plotMixtures.tidyLPA}
\title{Create latent profile plots}
\usage{
plotMixtures(x, variables = NULL, ci = 0.95, sd = TRUE,
  rawdata = TRUE, bw = FALSE, alpha_range = c(0, 0.1), ...)

plot_profiles(x, variables = NULL, ci = 0.95, sd = TRUE,
  rawdata = TRUE, bw = FALSE, alpha_range = c(0, 0.1), ...)

\method{plotMixtures}{default}(x, variables = NULL, ci = 0.95,
  sd = TRUE, rawdata = TRUE, bw = FALSE, alpha_range = c(0, 0.1),
  ...)

\method{plotMixtures}{tidyLPA}(x, variables = NULL, ci = 0.95,
  sd = TRUE, rawdata = TRUE, bw = FALSE, alpha_range = c(0, 0.1),
  ...)
}
\arguments{
\item{x}{An object containing the results of a mixture model analysis.}

\item{variables}{A character vectors with the names of the variables to be
plotted (optional).}

\item{ci}{Numeric. What confidence interval should the errorbars span? Defaults to
a 95\% confidence interval. Set to NULL to remove errorbars.}

\item{sd}{Logica. Whether to display a box encompassing +/- 1SD Defaults to
TRUE.}

\item{rawdata}{Should raw data be plotted in the background? Setting this to
TRUE might result in long plotting times.}

\item{bw}{Logical. Should the plot be black and white (for print), or color?}

\item{alpha_range}{The minimum and maximum values of alpha (transparancy) for
the raw data. Minimum should be 0; lower maximum values of alpha can help
reduce overplotting.}

\item{...}{Arguments passed to and from other functions.}
}
\value{
An object of class 'ggplot'.
}
\description{
Creates a profile plot according to best practices, focusing on the
visualization of classification uncertainty by showing:
\enumerate{
\item Bars reflecting a confidence interval for the class centroids
\item Boxes reflecting the standard deviations within each class; a box
encompasses +/- 64% of the observations in a normal distribution
\item Raw data, whose transparancy is weighted by the posterior class
probability, such that each datapoint is most clearly visible for the class
it is most likely to be a member of.
}
}
\examples{
\dontrun{
createMixtures(classes = 1:4, filename_stem = "cars",
               model_overall = "wt ON drat;",
               model_class_specific = "wt;  qsec;",
               rdata = mtcars,
               usevariables = c("wt", "qsec", "drat"),
               OUTPUT = "standardized")
runModels(replaceOutfile = "modifiedDate")
cars_results <- readModels(filefilter = "cars")
plotMixtures(cars_results, rawdata = TRUE)
}
\dontrun{
plotMixtures(cars_results, variables = "wt")
}
\dontrun{
plotMixtures(cars_results, coefficients = "stdyx.standardized")
}
}
\author{
Caspar J. van Lissa
}
\keyword{mixture}
\keyword{plot}
