% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{poms}
\alias{poms}
\title{Apply POMS-coding to data}
\usage{
poms(data)
}
\arguments{
\item{data}{A data.frame.}
}
\value{
A data.frame.
}
\description{
Takes in a data.frame, and applies POMS (proportion of of maximum)-coding to
the numeric columns.
}
\examples{
data <- data.frame(a = c(1, 2, 2, 4, 1, 6),
                   b = c(6, 6, 3, 5, 3, 4),
                   c = c("a", "b", "b", "t", "f", "g"))
poms(data)
}
\author{
Caspar J. van Lissa
}
