% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-solutions.R
\name{print.bestLPA}
\alias{print.bestLPA}
\title{Print bestLPA}
\usage{
\method{print}{bestLPA}(x, digits = getOption("digits"), na.print = "",
  ...)
}
\arguments{
\item{x}{An object of class 'bestLPA'.}

\item{digits}{minimal number of significant digits, see
\code{\link[base]{print.default}}.}

\item{na.print}{a character string which is used to indicate NA values in
printed output, or NULL. See \code{\link[base]{print.default}}.}

\item{...}{further arguments to be passed to or from other methods. They are
ignored in this function.}
}
\description{
S3 method 'print' for class 'bestLPA'.
}
\examples{
\dontrun{
if(interactive()){
results <- iris \%>\%
  select(Sepal.Length, Sepal.Width, Petal.Length, Petal.Width) \%>\%
  estimate_profiles(1:3) \%>\%
  compare_solutions()
print(results)
}
}
}
