% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-profiles.R
\name{print.tidyLPA}
\alias{print.tidyLPA}
\title{Print tidyLPA}
\usage{
\method{print}{tidyLPA}(x, stats = c("AIC", "BIC", "Entropy", "prob_min",
  "prob_max", "n_min", "n_max", "BLRT_p"), digits = 2, na.print = "",
  ...)
}
\arguments{
\item{x}{An object of class 'tidyLPA'.}

\item{stats}{Character vector. Statistics to be printed. Default:
c("AIC", "BIC", "Entropy", "prob_min", "prob_max", "n_min", "n_max", "BLRT_p"
).}

\item{digits}{minimal number of significant digits, see
\code{\link[base]{print.default}}.}

\item{na.print}{a character string which is used to indicate NA values in
printed output, or NULL. See \code{\link[base]{print.default}}.}

\item{...}{further arguments to be passed to or from other methods. They are
ignored in this function.}
}
\description{
S3 method 'print' for class 'tidyLPA'.
}
\examples{
\dontrun{
if(interactive()){
 iris \%>\%
   select(Sepal.Length, Sepal.Width, Petal.Length, Petal.Width) \%>\%
   estimate_profiles(3)
 }
}
}
\author{
Caspar J. van Lissa
}
