% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-profiles.R
\name{print.tidyProfile}
\alias{print.tidyProfile}
\title{Print tidyProfile}
\usage{
\method{print}{tidyProfile}(x, digits = 3, na.print = "", ...)
}
\arguments{
\item{x}{An object of class 'tidyProfile'.}

\item{digits}{minimal number of significant digits, see
\code{\link[base]{print.default}}.}

\item{na.print}{a character string which is used to indicate NA values in
printed output, or NULL. See \code{\link[base]{print.default}}.}

\item{...}{further arguments to be passed to or from other methods. They are
ignored in this function.}
}
\description{
S3 method 'print' for class 'tidyProfile'.
}
\examples{
\dontrun{
if(interactive()){
 tmp <- iris \%>\%
   select(Sepal.Length, Sepal.Width, Petal.Length, Petal.Width) \%>\%
   estimate_profiles(3)
 tmp[[2]]
 }
}
}
\author{
Caspar J. van Lissa
}
