% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_models_mplus.R
\name{select_ancillary_functions}
\alias{select_ancillary_functions}
\title{Create models for a specific mclust model}
\usage{
select_ancillary_functions(df, ...)
}
\arguments{
\item{df}{data.frame with two or more columns with continuous variables}

\item{...}{unquoted variable names separated by commas}

\item{data_filename}{name of data file to prepare; defaults to d.dat}

\item{script_filename}{name of script to prepare; defaults to t.imp}

\item{the_title}{title of the model; defaults to test}
}
\value{
either a tibble or a ggplot2 plot of the BIC values for the explored models
}
\description{
Create models for a specific mclust model
}
\details{
Creates profiles (or estimates of the mixture components) for a specific mclust model in terms of the specific number of mixture components and the structure of the residual covariance matrix
}
\examples{
d <- pisaUSA15
d <- dplyr::sample_n(d, 200)
m3 <- create_profiles_lpa(d,
                          broad_interest, enjoyment, self_efficacy,
                          n_profiles = 3, to_return="tibble")
}
