% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary_mclust}
\alias{summary_mclust}
\title{Calculate summary statistics for mclust output}
\usage{
summary_mclust(x)
}
\arguments{
\item{x}{output from create_profiles_mclust() when to_return is set to TRUE}
}
\description{
Calculate summary statistics for mclust output
}
\details{
groups by profile and calculates the mean
}
