% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_centroids_mclust}
\alias{calculate_centroids_mclust}
\title{Calculate centroids from an mclust model object}
\usage{
calculate_centroids_mclust(x)
}
\arguments{
\item{x}{an object of class `Mclust`}
}
\description{
Calculate centroids from an mclust model object
}
\details{
Extract the output of the mclust output from the function create_profiles() so that posterior probabilities for specific observations, statistics related to the estimation, and other output can be viewed
}
