% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_profiles_mclust}
\alias{create_profiles_mclust}
\title{Create profiles for a specific mclust model}
\usage{
create_profiles_mclust(df, n_profiles, variance_structure = "freed",
  covariance_structure = "freed", model_name = NULL, to_return = "tibble")
}
\arguments{
\item{df}{data.frame with two or more columns with continuous variables}

\item{n_profiles}{the number of profiles (or mixture components) to be estimated}

\item{model_name}{the mclust model to explore: "means", "means_varying_covariance", and "means_varying_covariance" specify the three most common models, in order from most to least constrained; run ?mclust::mclustModelNames() to see all of the possible models and their names / abbreviations)}

\item{to_return}{character string for whether to return a tibble or the mclust output; if a tibble is returned, the mclust output can be viewed using the extract_mclust_output() function, with the tibble as its only argument}
}
\value{
a ggplot2 plot of the BIC values for the explored models
}
\description{
Create profiles for a specific mclust model
}
\details{
Creates profiles (or estimates the mixture components) for a specific mclust model in terms of the specific number of mixture components and the structure of the residual covariance matrix
}
\examples{
library(dplyr)
df <- select(iris, -Species)
create_profiles_mclust(df, n_profiles = 3, variance_structure = "freed", covariance_structure = "freed")
}
