% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_models_mplus.R
\name{create_profiles_mplus}
\alias{create_profiles_mplus}
\title{Create models for a specific mclust model}
\usage{
create_profiles_mplus(df, ..., n_profiles, the_title = "test",
  data_filename = "d.dat", script_filename = "t.inp",
  output_filename = "t.out", model = 1, start_iterations = c(100, 20))
}
\arguments{
\item{df}{data.frame with two or more columns with continuous variables}

\item{...}{unquoted variable names separated by commas}

\item{n_profiles}{the number of profiles (or mixture components) to be estimated}

\item{the_title}{title of the model; defaults to test}

\item{data_filename}{name of data file to prepare; defaults to d.dat}

\item{script_filename}{name of script to prepare; defaults to t.inp}

\item{output_filename}{name of the output; defaults to t.out}

\item{model}{the mclust model to explore: 1 (varying means, equal variances, and residual covariances fixed to 0); 2 (varying means, equal variances and covariances; 3 (varying means and variances, covariances fixed to 0), 4 (varying means and covariances, equal variances; can only be specified in Mplus); and 5 (varying means, variances, and covariances), in order least to most freely-estimated; see the introductory vignette for more information}

\item{start_iterations}{the number of start iterations; defaults to c(100, 20)}
}
\value{
either a tibble or a ggplot2 plot of the BIC values for the explored models
}
\description{
Create models for a specific mclust model
}
\details{
Creates an mplus model (.inp) and associated data file (.dat)
}
\examples{
d <- tbl_df(iris[, -5])
m1 <- create_profiles_mplus(d,
                            Sepal.Length, Sepal.Width, Petal.Length, Petal.Width,
                            n_profiles = 2,
                            model = 1)
}
