% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{explore_models_mclust}
\alias{explore_models_mclust}
\title{Explore BIC of mclust models}
\usage{
explore_models_mclust(df, ..., n_profiles_range = 1:9,
  model_names = c("EEI", "EEE", "VVV"), statistic = "BIC",
  return_table = FALSE)
}
\arguments{
\item{df}{data.frame with two or more columns with continuous variables}

\item{...}{unquoted variable names separated by commas}

\item{n_profiles_range}{a vector with the range of the number of mixture components to explore; defaults to 1 through 9 (1:9)}

\item{model_names}{mclust models to explore; defaults to constrained variance, fixed variances ("EII"), constrained variance, constrained covariance ("EEE"), and freed variance, freed covariance ("VVV"); run mclustModelNames() from mclust to see all of the possible models and their names / abbreviations}

\item{statistic}{what statistic to plot; BIC or ICL are presently available as options}

\item{return_table}{logical (TRUE or FALSE) for whether to return a table of the output intsead of a plot; defaults to FALSE}
}
\value{
a ggplot2 plot of the BIC values for the explored models
}
\description{
Explore BIC of mclust models
}
\details{
Explore the BIC values of a range of models in terms of a) the structure of the residual covariance matrix and b) the number of mixture components (or profiles)
}
\examples{
\dontrun{
   d <- pisaUSA15
   explore_models_mclust(d, broad_interest, instrumental_mot, self_efficacy)
}
}
