% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_covariance}
\alias{extract_covariance}
\title{Extract mclust covariance}
\usage{
extract_covariance(x, profile_n)
}
\arguments{
\item{x}{an object of class `Mclust`}

\item{profile_n}{the number of profiles}
}
\description{
Extract mclust covariance
}
\details{
Extract the covariances
}
