% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_mclust.R
\name{extract_covariance}
\alias{extract_covariance}
\title{Extract mclust covariances}
\usage{
extract_covariance(x)
}
\arguments{
\item{x}{object of class Mclust}
}
\description{
Extract mclust covariances
}
\details{
Extract the covariances
}
\examples{
d <- pisaUSA15
d <- sample_n(pisaUSA15, 500)
m <- create_profiles_lpa(d, broad_interest, enjoyment, self_efficacy, n_profiles = 4, to_return = "mclust")
extract_covariance(m)
}
