% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_mclust.R
\name{extract_mclust_summary}
\alias{extract_mclust_summary}
\title{Extract mclust summary statistics}
\usage{
extract_mclust_summary(x)
}
\arguments{
\item{x}{object of class Mclust}
}
\description{
Extract mclust summary statistics
}
\details{
Extract the log likelihood, BIC, and entropy statistics
}
