% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_mclust.R
\name{extract_variance}
\alias{extract_variance}
\title{Extract mclust variances}
\usage{
extract_variance(x)
}
\arguments{
\item{x}{object of class Mclust}
}
\description{
Extract mclust variances
}
\details{
Extract the variances and covariances
}
\examples{
d <- pisaUSA15
d <- sample_n(pisaUSA15, 500)
m <- create_profiles_lpa(d, broad_interest, enjoyment, self_efficacy, n_profiles = 4, to_return = "mclust")
extract_variance(m)
}
