% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_variance}
\alias{extract_variance}
\title{Extract mclust variances}
\usage{
extract_variance(x, profile_n)
}
\arguments{
\item{x}{an object of class `Mclust`}

\item{profile_n}{the number of profiles}
}
\description{
Extract mclust variances
}
\details{
Extract the variances and covariances
}
