% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profiles.R
\name{plot_profiles_lpa}
\alias{plot_profiles_lpa}
\title{Plot profile centroids}
\usage{
plot_profiles_lpa(d, to_center = F, to_scale = F, plot_what = "tibble")
}
\arguments{
\item{d}{output from create_profiles_mclust()}

\item{to_center}{whether to center the data before plotting}

\item{to_scale}{whether to scale the data before plotting}

\item{plot_what}{whether to plot tibble or mclust output from create_profiles_lpa(); defaults to tibble}
}
\description{
Plot profile centroids
}
\details{
Plot the centroids for tibble or mclust output from create_profiles_lpa()
}
