% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profiles_mplus.R
\name{plot_profiles_mplus}
\alias{plot_profiles_mplus}
\title{Plot profile centroids for mplus output}
\usage{
plot_profiles_mplus(mplus_data, to_center = T, to_scale = T)
}
\arguments{
\item{mplus_data}{output from estimate_profiles_mclust() with return_savedata = T specified}

\item{to_center}{whether to center the data before plotting}

\item{to_scale}{whether to scale the data before plotting}
}
\description{
Plot profile centroids for mplus output
}
\details{
Plot the centroids for tibble or mclust output from estimate_profiles_mplus()
}
